+++ src/extension/utils.ts
@@ -22,7 +22,6 @@
  */
 /* eslint-disable no-console */
 
-import {URLReader} from '@malloydata/malloy';
 import {BaseLanguageClient} from 'vscode-languageclient';
 import * as vscode from 'vscode';
 import {
@@ -30,6 +29,7 @@ import {
   FetchCellDataEvent,
   FetchFileEvent,
   CellData,
+  FileHandler,
 } from '../common/types';
 
 /**
@@ -115,39 +115,44 @@ export async function fetchCellData(uriString: string): Promise<CellData[]> {
   return result;
 }
 
-export class VSCodeURLReader implements URLReader {
+export class ClientFileHandler implements FileHandler {
+  /**
+   * Requests a file from the worker's controller. Although the
+   * file path is a file system path, reading the file off
+   * disk doesn't take into account unsaved changes that only
+   * VS Code is aware of.
+   *
+   * @param uri URI to resolve
+   * @returns File contents
+   */
+  async fetchFile(uri: string): Promise<string> {
+    return fetchFile(uri);
+  }
+
+  /**
+   * Requests a binary file from the worker's controller.
+   *
+   * @param uri URI to resolve
+   * @returns File contents
+   */
+
+  async fetchBinaryFile(uri: string): Promise<Uint8Array> {
+    return fetchBinaryFile(uri);
+  }
+
+  /**
+   * Requests a set of cell data from the worker's controller.
+   *
+   * @param uri URI to resolve
+   * @returns File contents
+   */
+  async fetchCellData(uri: string): Promise<CellData[]> {
+    return fetchCellData(uri);
+  }
+
   async readURL(url: URL): Promise<string> {
-    return fetchFile(url.toString());
+    return this.fetchFile(url.toString());
   }
 }
 
-export const initFileMessaging = (
-  context: vscode.ExtensionContext,
-  client: BaseLanguageClient
-) => {
-  context.subscriptions.push(
-    client.onRequest('malloy/fetchFile', async (event: FetchFileEvent) => {
-      console.info('fetchFile returning', event.uri);
-      return await fetchFile(event.uri);
-    })
-  );
-
-  context.subscriptions.push(
-    client.onRequest(
-      'malloy/fetchBinaryFile',
-      async (event: FetchBinaryFileEvent) => {
-        console.info('fetchBinaryFile returning', event.uri);
-        return await fetchBinaryFile(event.uri);
-      }
-    )
-  );
-  context.subscriptions.push(
-    client.onRequest(
-      'malloy/fetchCellData',
-      async (event: FetchCellDataEvent) => {
-        console.info('fetchCellData returning', event.uri);
-        return await fetchCellData(event.uri);
-      }
-    )
-  );
-};
+export const fileHandler = new ClientFileHandler();
