+++ src/worker/file_handler.ts
@@ -22,37 +22,7 @@
  */
 
 import * as rpc from 'vscode-jsonrpc';
-import {URLReader} from '@malloydata/malloy';
-import {CellData} from '../common/types';
-
-export interface FileHandler extends URLReader {
-  /**
-   * Requests a file from the worker's controller. Although the
-   * file path is a file system path, reading the file off
-   * disk doesn't take into account unsaved changes that only
-   * VS Code is aware of.
-   *
-   * @param uri URI to resolve
-   * @returns File contents
-   */
-  fetchFile(uri: string): Promise<string>;
-  /**
-   * Requests a binary file from the worker's controller.
-   *
-   * @param uri URI to resolve
-   * @returns File contents
-   */
-
-  fetchBinaryFile(uri: string): Promise<Uint8Array>;
-
-  /**
-   * Requests a set of cell data from the worker's controller.
-   *
-   * @param uri URI to resolve
-   * @returns File contents
-   */
-  fetchCellData(uri: string): Promise<CellData[]>;
-}
+import {CellData, FileHandler} from '../common/types';
 
 export class RpcFileHandler implements FileHandler {
   constructor(private connection: rpc.MessageConnection) {}
